% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AUC_WD_metric.R
\name{AUC_WD_metric}
\alias{AUC_WD_metric}
\title{AUC and WD function}
\usage{
AUC_WD_metric(alpha, delta, K, R)
}
\arguments{
\item{alpha}{The first shape parameter estimated for the beta mixture model.}

\item{delta}{The second shape parameter estimated for the beta mixture model.}

\item{K}{The number of clusters estimated.}

\item{R}{The number of sample types in the dataset.}
}
\value{
The list with AUC and WD values.
}
\description{
Function to find the level of similarities between the \eqn{R} cumulative distributions estimated in each of the \eqn{K} clusters.
}
\details{
Function to find the level of similarities between the \eqn{R} cumulative distributions estimated in each of the \eqn{K} clusters.
}
\seealso{
\code{\link{betaclust}}
}
