% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_c.R
\name{beta_c}
\alias{beta_c}
\title{C.. Model from the family of beta mixture models for DNA methylation data}
\usage{
beta_c(data, K = 3, seed, register = NULL)
}
\arguments{
\item{K}{number of methylation groups to be identified (default=3)}

\item{seed}{seed for reproducible work}

\item{register}{setting for parallelization}

\item{X}{methylation values for CpG sites frpm R samples collected from N patients}
}
\description{
C.. Model from the family of beta mixture models for DNA methylation data
}
