% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_kn.R
\name{beta_kn}
\alias{beta_kn}
\title{The KN. model}
\usage{
beta_kn(data, M = 3, seed = NULL)
}
\arguments{
\item{data}{A dataframe of dimension \eqn{C*N} containing methylation values for \eqn{C} CpG sites from \eqn{R = 1} samples collected from \eqn{N} patients.
Samples are grouped together in the dataframe such that the columns are ordered as Sample1_Patient1, Sample1_Patient2, etc.}

\item{M}{Number of methylation states to be identified in a DNA sample.}

\item{seed}{Seed to allow for reproducibility (default = NULL).}
}
\value{
A list containing:
\itemize{
   \item cluster_size - the total number of CpG sites in each of the K clusters.
   \item llk - a vector containing the log-likelihood value at each step of the EM algorithm.
   \item data - this contains the methylation dataset along with the cluster label for each CpG site.
   \item alpha - this contains the first shape parameter for the beta mixture model.
   \item delta - this contains the second shape parameter for the mixture model.
   \item tau - the proportion of CpG sites in each cluster.
   \item z - a matrix of dimension \eqn{C*K} containing the posterior probability of each CpG site belonging to each of the \eqn{K} clusters.
   \item uncertainty - the uncertainty of each CpG site's clustering.    }
}
\description{
Fit the KN. model from the family of beta mixture models for DNA methylation data.
             The KN. model analyses a single DNA sample and identifies the thresholds between the different methylation states.
}
\details{
The KN. model clusters each of the \eqn{C} CpG sites into one of \eqn{K} methylation states, based on data from \eqn{N} patients for one DNA sample (i.e. \eqn{R = 1}).
As each CpG site can belong to either of the \eqn{M = 3} methylation states (hypomethylated, hemimethylated or hypermethylated), the default value of \eqn{K = M = 3}.
The KN. model differs from the C.. model as it is less parsimonious, allowing cluster and patient-specific shape parameters. The returned object from this function can be passed as an input parameter to the
\code{\link[betaclust:threshold]{threshold}} function available in this package to calculate the thresholds between the methylation states.
}
\examples{
\dontrun{
data(pca.methylation.data)
my.seed = 190
M = 3
data_output = beta_kn(pca.methylation.data[,2:5],M,seed = my.seed)
thresholds = threshold(data_output,"KN.")
}
}
\seealso{
\code{\link{beta_k}}

\code{\link{betaclust}}

\code{\link{threshold}}
}
