% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_kr.R
\name{beta_kr}
\alias{beta_kr}
\title{The K.R Model}
\usage{
beta_kr(data, M = 3, N, R, seed = NULL)
}
\arguments{
\item{data}{A dataframe of dimension \eqn{C*NR} containing methylation values for \eqn{C} CpG sites from \eqn{R} samples collected from \eqn{N} patients.
Samples are grouped together in the dataframe such that the columns are ordered as Sample1_Patient1, Sample1_Patient2, Sample2_Patient1, Sample2_Patient2, etc.}

\item{M}{Number of methylation states to be identified.}

\item{N}{Number of patients in the study.}

\item{R}{Number of samples collected from each patient for study.}

\item{seed}{Seed to allow for reproducibility (default = NULL).}
}
\value{
A list containing:
\itemize{
   \item cluster_size - the total number of CpG sites in each of the K clusters.
   \item llk - a vector containing the log-likelihood value at each step of the EM algorithm.
   \item data - this contains the methylation dataset along with the cluster label for each CpG site.
   \item alpha - this contains the first shape parameter for the beta mixture model.
   \item delta - this contains the second shape parameter for the beta mixture model.
   \item tau - the proportion of CpG sites in each cluster.
   \item z - a matrix of dimension \eqn{C*K} containing the posterior probability of each CpG site belonging to each of the \eqn{K} clusters.
   \item uncertainty - the uncertainty of each CpG site's clustering.    }
}
\description{
A beta mixture model for identifying differentially methylated CpG sites between \eqn{R} DNA samples collected from \eqn{N} patients.
}
\details{
The K.R model allows identification of the differentially methylated CpG sites between the \eqn{R} DNA samples collected from each of \eqn{N} patients.
As each CpG site in a DNA sample can belong to one of \eqn{M} methylation states, there can be \eqn{K=M^R} methylation state changes between \eqn{R} DNA samples.
The shape parameters vary for each DNA sample but are constrained to be equal for each patient. An initial clustering using k-means is performed to identify \eqn{K} clusters. The resulting clustering solution is provided as
starting values to the Expectation-Maximisation algorithm. A digamma approximation is used to obtain the maximised
parameters in the M-step instead of a computationally inefficient numerical optimisation step.
}
\examples{
\dontrun{
data(pca.methylation.data)
my.seed = 190
M = 3
N = 4
R = 2
data_output = beta_kr(pca.methylation.data[,2:5],M,N,R,seed = my.seed)
}
}
\seealso{
\code{\link{betaclust}}
}
