% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecdf.betaclust.R
\name{ecdf.betaclust}
\alias{ecdf.betaclust}
\title{The empirical cumulative distribution function}
\usage{
ecdf.betaclust(
  x,
  samples = 2,
  sample_name = c("Sample 1", "Sample 2"),
  title = NULL
)
}
\arguments{
\item{x}{A dataframe containing methylation values of identified differentially methylated regions related to a gene.
Group each sample together in the dataframe such that the columns are ordered as Sample1_Patient1, Sample1_Patient2, Sample2_Patient1, Sample2_Patient2.}

\item{samples}{number of tissue samples from which DNA methylation data are collected (default samples = 2).}

\item{sample_name}{The order in which the samples are grouped in the dataframe x (default = c("Sample 1","Sample 2")).}

\item{title}{The title that the user wants to display on the graph. If no title is to be displayed the default is "NULL" value.}
}
\value{
The ECDF plot for the selected CpG sites for all patients and their DNA samples.
}
\description{
An empirical cumulative distribution function (ECDF) plot for a betaclust object.
}
\details{
This function plots the ECDF graphs of the differentially methylated CpG sites identified using the K.R model for all patient samples.
The graph visualises the methylation profile changes between the different DNA samples for each patient.
}
\seealso{
\code{\link{betaclust}}

\code{\link{beta_kr}}
}
