% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_aic.R
\name{em_aic}
\alias{em_aic}
\title{Akaike Information Criterion}
\usage{
em_aic(llk, C, K, patients = 4, samples = 1, model_names = "C..")
}
\arguments{
\item{llk}{log-likelihood value}

\item{C}{number of CpG sites}

\item{K}{number of clusters}

\item{patients}{number of patients}

\item{samples}{no. of samples}

\item{model_names}{mixture model (method=c("C..","CN.","C.R"))}
}
\value{
The AIC value for the selected model.
}
\description{
The AIC value used to select the optimal model.
}
\details{
Computes the AIC for the beta mixture models given the loglikelihood, the dimension of the data, and the mixture model names.
}
\seealso{
\code{\link{em_bic}}

\code{\link{em_icl}}
}
