% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_bic.R
\name{em_bic}
\alias{em_bic}
\title{Bayesian Information Criterion}
\usage{
em_bic(llk, C, K, patients = 4, samples = 1, model_names = "C..")
}
\arguments{
\item{llk}{log-likelihood value}

\item{C}{number of CpG sites}

\item{K}{number of clusters}

\item{patients}{number of patients}

\item{samples}{no. of samples}

\item{model_names}{mixture model (method=c("C..","CN.","C.R"))}
}
\value{
The BIC value for the selected model
}
\description{
The BIC value used to select the optimal model
}
