% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_icl.R
\name{em_icl}
\alias{em_icl}
\title{Integrated Complete-data Likelihood (ICL) Criterion}
\usage{
em_icl(llk, C, M, patients = 4, samples = 1, model_name = "K..", z)
}
\arguments{
\item{llk}{log-likelihood value.}

\item{C}{number of CpG sites.}

\item{M}{number of methylation states identified in a DNA sample.}

\item{patients}{number of patients.}

\item{samples}{number of DNA samples collected from each patient.}

\item{model_name}{fitted mixture model (method = c("K..","KN.","K.R")).}

\item{z}{matrix containing the probability for each CpG site of belonging to each of the K clusters.}
}
\value{
The ICL value for the selected model.
}
\description{
Compute the ICL value for the optimal model.
}
\details{
Computes the ICL for a specified model given the log-likelihood, the dimension of the data, and the model specification. This criterion penalises the BIC by including the entropy term favouring the
well separated clusters.
}
\seealso{
\code{\link{em_aic}}

\code{\link{em_bic}}
}
