% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy.data.R
\docType{data}
\name{legacy.data}
\alias{legacy.data}
\title{MethylationEPIC manifest data.}
\format{
A data frame with 867525 rows and 6 columns.
\describe{
  \item{IlmnID}{This contains the unique identifier from the Illumina CG database, i.e. the probe ID.}
  \item{Genome_Build}{Genome build referenced by the Infinium MethylationEPIC manifest.}
  \item{CHR}{Chromosome containing the CpG (Genome_Build = 37).}
  \item{MAPINFO}{This contains the methylation values from benign prostate tissue collected from patient 3.}
  \item{UCSC_RefGene_Name}{Target gene name(s), from the UCSC database.
  Note: multiple listings of the same gene name indicate splice variants.}
  \item{UCSC_CpG_Islands_Name}{Chromosomal coordinates of the CpG Island from UCSC.}
   }
}
\usage{
data(legacy.data)
}
\description{
The dataset contains the manifest data from the Illumina MethylationEPIC beadchip array.
}
\seealso{
\code{\link{pca.methylation.data}}
}
\keyword{datasets}
