% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.methylation.data.R
\docType{data}
\name{pca.methylation.data}
\alias{pca.methylation.data}
\title{DNA methylation dataset of patients suffering from prostate cancer disease.}
\format{
A data frame with 694820 rows and 9 columns. The data contains no missing values.
\describe{
  \item{IlmnID}{This contains the Unique identifier from the Illumina CG database. (The probe ID).}
  \item{Patient_benign_1}{This contains the methylation values from benign prostate tissue collected from patient 1.}
  \item{Patient_benign_2}{This contains the methylation values from benign prostate tissue collected from patient 2.}
  \item{Patient_benign_3}{This contains the methylation values from benign prostate tissue collected from patient 3.}
  \item{Patient_benign_4}{This contains the methylation values from benign prostate tissue collected from patient 4.}
  \item{Patient_benign_1}{This contains the methylation values from tumor prostate tissue collected from patient 1.}
  \item{Patient_benign_2}{This contains the methylation values from tumor prostate tissue collected from patient 2.}
  \item{Patient_benign_3}{This contains the methylation values from tumor prostate tissue collected from patient 3.}
  \item{Patient_benign_4}{This contains the methylation values from tumor prostate tissue collected from patient 4.}
   }
}
\usage{
data(pca.methylation.data)
}
\description{
The dataset contains pre-processed beta methylation values of R=2 samples which are collected from N=4 patients suffering from prostate cancer disease.
}
\keyword{datasets}
