% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.betaclust.R, R/print.summary.betaclust.R
\name{print.summary.betaclust}
\alias{print.summary.betaclust}
\title{Plots for betaclust object}
\usage{
\method{print}{summary.betaclust}(x, digits = getOption("digits"), ...)

\method{print}{summary.betaclust}(x, digits = getOption("digits"), ...)
}
\arguments{
\item{x}{summary of betaclust object}

\item{object}{betaclust object}

\item{what}{The different plots that can be obtained from the object (default="density") (what=c("density","uncertainty","InformationCriterion"))}

\item{plot_type}{The plot type to be displayed (default="ggplot")(plot_type=c("ggplot","plotly"))}
}
\description{
Plots for betaclust object

Print summary of betaclust output
}
