% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.betaclust.R
\name{summary.betaclust}
\alias{summary.betaclust}
\title{Summary statistics of betaclust output}
\usage{
\method{summary}{betaclust}(object)
}
\arguments{
\item{x}{betaclust object}
}
\value{
An object of class "summary.betaclust".
The object returns the following list of values:
\itemize{
\item CpG_sites - The number of CpG sites analysed using the beta mixture models.
\item patients - The number of patients analysed using the beta mixture models.
\item samples - The numder of samples analysed using the beta mixture models.
\item cluster_count - The number of groups, the data is clustered into.
\item modelName - The optimal model selected.
\item loglik - The log-likelihood value for the selected optimal model.
\item Information_criterion - The information criterion used to select the optimal model.
\item ic_output - This stores the information criterion value calculated for each model.
\item classification - The total number of CpG sites identified in each cluster. }
}
\description{
Calculates and prints the summary statistics of the optimal model selected for printing.
}
