% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold.R
\name{threshold}
\alias{threshold}
\title{Thresholds for K.. and KN. models}
\usage{
threshold(object, model_name)
}
\arguments{
\item{object}{A dataframe containing methylation values of identified differentially methylated regions related to a gene.}
}
\value{
The thresholds calculated for the clustering solution
}
\description{
An empirical cumulative distribution function (ECDF) plot for a betaclust object.
}
\details{
This function plots the ECDF graphs of the differentially methylated CpG sites identified using the K.R model for all patient samples.
The graph visualises the methylation profile changes between the different DNA samples for each patient.
}
\seealso{
\code{\link{betaclust}}

\code{\link{beta_k}}

\code{\link{beta_kn}}
}
