% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_cn.R
\name{beta_cn}
\alias{beta_cn}
\title{The CN. model}
\usage{
beta_cn(data, K = 3, seed, register = NULL)
}
\arguments{
\item{K}{number of methylation groups to be identified (default=3)}

\item{seed}{seed for reproducible work}

\item{register}{setting for parallelization}

\item{X}{methylation values for CpG sites frpm R samples collected from N patients}
}
\value{
A list of clustering solution results.
\itemize{
   \item cluster_count - The total number of CpG sites identified in each cluster.
   \item llk - The vector containing log-likelihood values calculated for each step of parameter estimation.
   \item data - This contains the methylation dataset along with the cluster label as determined by the mixture model.
   \item alpha - This contains the shape parameter 1 for the beta mixtures for \eqn{K^R} groups.
   \item beta - This contains the shape parameter 2 for the beta mixtures for \eqn{K^R} groups.
   \item tau - The proportion of CpG sites in each cluster.
   \item z - The matrix contains the probability calculated for each CpG site belonging to the \eqn{K^R} clusters.
   \item uncertainty - The uncertainty of a CpG site belonging to the identified cluster.
   }
}
\description{
The CN. model from the family of beta mixture models for DNA methylation data.
             This model analyses a single DNA sample and identify the thresholds for the different methylation profiles.
}
\details{
This model clusters each of the C CpG sites into one of K = M methylation states, based on data from N patients where R = 1.
The default value for M = 3 as a CpG site can be either hypomethylated, hemimethylated or hypermethylated.
The CN. model differs from the C.. model as it is less parsimonious, allowing cluster and patient-specific shape parameters.
}
\examples{
\dontrun{
data(pca.methylation.data)
my.seed=190
K=3
data_output=beta_cn(pca.methylation.data[,2:5],K,seed=my.seed)
}
}
\seealso{
\code{\link{beta_c}}

\code{\link{betaclust}}
}
