% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_cn.R
\name{beta_cn}
\alias{beta_cn}
\title{The CN. model}
\usage{
beta_cn(data, K = 3, seed, register = NULL)
}
\arguments{
\item{K}{number of methylation groups to be identified (default=3)}

\item{seed}{seed for reproducible work}

\item{register}{setting for parallelization}

\item{X}{methylation values for CpG sites frpm R samples collected from N patients}
}
\value{
A list of clustering solution results.
\itemize{
   \item cluster_count - The total number of CpG sites identified in each cluster.
   \item llk - The vector containing log-likelihood values calculated for each step of parameter estimation.
   \item data - This contains the methylation dataset along with the cluster label as determined by the mixture model.
   \item alpha - This contains the shape parameter 1 for the beta mixtures for K^R groups.
   \item beta - This contains the shape parameter 2 for the beta mixtures for K^R groups.
   \item tau - The proportion of CpG sites in each cluster.
   \item z - The matrix contains the probability calculated for each CpG site belonging to the K^R clusters.
   \item uncertainty - The uncertainty of a CpG site belonging to the identified cluster.
   }
}
\description{
CN. Model from the family of beta mixture models for DNA methylation data.
             This model analyses a single DNA sample collected from N patients to cluster the
             CpG sites into K groups. By default K=3 (hypomethylation, hemimethylation and hypermethylation).
}
\examples{
\dontrun{
data(pca.methylation.data)
my.seed=190
K=3
data_output=beta_cn(pca.methylation.data[,2:5],K,seed=my.seed)
}
}
