% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_cn.R
\name{beta_cn}
\alias{beta_cn}
\title{CN. Model from the family of beta mixture models for DNA methylation data}
\usage{
beta_cn(X, K = 3, seed, register = NULL)
}
\arguments{
\item{X}{methylation values for CpG sites frpm R samples collected from N patients}

\item{K}{number of methylation groups to be identified (default=3)}

\item{seed}{seed for reproducible work}

\item{register}{setting for parallelization}
}
\description{
CN. Model from the family of beta mixture models for DNA methylation data
}
