% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_cr.R
\name{beta_cr}
\alias{beta_cr}
\title{The C.R Model}
\usage{
beta_cr(data, K = 3, patients, samples, seed, register = NULL)
}
\arguments{
\item{K}{number of methylation groups to be identified (default=3)}

\item{patients}{number of patients in the study}

\item{samples}{number of samples collected from each patient for study}

\item{seed}{seed for reproducible work}

\item{register}{setting for parallelization}

\item{X}{methylation values for CpG sites frpm R samples collected from N patients}
}
\value{
The model returns the maximised parameters and the log-likelihood value for them. The complete data with predicted cluster
membership and the uncertainty in the clustering solution is returned.
}
\description{
Beta mixture model for identifying differentially methylated regions between multiple DNA samples
}
\details{
An initial clustering using K-means is performed to identify K^samples cluster. These values are provided as
starting values to the Expectation-Maximisation algorithm. A digamma approximation is used to obtain the maximised
parameters in the M-step instead of the computationally inefficient numerical optimisation step.
}
\examples{
a+b
}
