% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_cr.R
\name{beta_cr}
\alias{beta_cr}
\title{C.R Model from the family of beta mixture models for DNA methylation data}
\usage{
beta_cr(X, K = 3, patients, samples, seed, register = NULL)
}
\arguments{
\item{X}{methylation values for CpG sites frpm R samples collected from N patients}

\item{K}{number of methylation groups to be identified (default=3)}

\item{patients}{number of patients in the study}

\item{samples}{number of samples collected from each patient for study}

\item{seed}{seed for reproducible work}

\item{register}{setting for parallelization}
}
\description{
C.R Model from the family of beta mixture models for DNA methylation data
}
