% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_kr.R
\name{beta_kr}
\alias{beta_kr}
\title{The K.R Model}
\usage{
beta_kr(data, M = 3, patients, samples, seed = NULL, register = NULL)
}
\arguments{
\item{data}{Methylation values for \eqn{C} CpG sites from \eqn{R} samples collected from \eqn{N} patients.}

\item{M}{Number of methylation states to be identified.}

\item{patients}{Number of patients in the study.}

\item{samples}{Number of samples collected from each patient for study.}

\item{seed}{Seed to allow for reproducibility.}

\item{register}{Setting for registering the parallel backend with the "foreach" package. To start parallel execution of R code on machine with multiple cores, "NULL" value needs to be assigned to this parameter.}
}
\value{
A list containing:
\itemize{
   \item cluster_size - the total number of CpG sites identified in each of the K clusters.
   \item llk - a vector containing the log-likelihood value at each step of the EM algorithm.
   \item data - this contains the methylation dataset along with the cluster label for each CpG site.
   \item alpha - this contains the first shape parameter for the beta mixture model.
   \item delta - this contains the second shape parameter for the beta mixture model.
   \item tau - the proportion of CpG sites in each cluster.
   \item z - a matrix containing the probability for each CpG site of belonging to each of the \eqn{K} clusters.
   \item uncertainty - the uncertainty of each CpG site's clustering.    }
}
\description{
A beta mixture model for identifying differentially methylated CpG sites between \eqn{R} DNA samples collected from \eqn{N} patients.
}
\details{
The K.R model allows identification of the differentially methylated CpG sites between the \eqn{R} DNA samples collected from each of \eqn{N} patients.
As each CpG site in a DNA sample can belong to either of \eqn{M} methylation states, there can be \eqn{K=M^R} methylation state changes between \eqn{R} DNA samples.
The parameters vary for each DNA sample but are constrained to be equal for each patient. An initial clustering using K-means is performed to identify \eqn{K} clusters. The resulting clustering solution is provided as
starting values to the Expectation-Maximisation algorithm. A digamma approximation is used to obtain the maximised
parameters in the M-step instead of a computationally inefficient numerical optimisation step.
}
\examples{
\dontrun{
data(pca.methylation.data)
my.seed = 190
M = 3
patients = 4
samples = 2
data_output = beta_kr(pca.methylation.data[,2:5],M,patients,samples,seed = my.seed)
}
}
\references{
{Microsoft, Weston, S. (2022): foreach: Provides Foreach Looping Construct. R package version 1.5.2. https://CRAN.R-project.org/package=foreach.}
}
\seealso{
\code{\link{betaclust}}
}
