% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaclust.R
\name{betaclust}
\alias{betaclust}
\title{The betaclust wrapper function}
\usage{
betaclust(
  data,
  M = 3,
  patients,
  samples,
  model_names = "K..",
  model_selection = "BIC",
  seed,
  register = NULL
)
}
\arguments{
\item{data}{Methylation values for \eqn{C} CpG sites from \eqn{R} samples collected from \eqn{N} patients.}

\item{M}{Number of methylation states to be identified in a DNA sample.}

\item{patients}{Number of patients in the study.}

\item{samples}{Number of samples collected from each patient for the study.}

\item{model_names}{Models to run from the set of models, K.., KN. and K.R, default = K.. . See details.}

\item{model_selection}{Information criterion used for model selection. Options are AIC/BIC/ICL/default = BIC.}

\item{seed}{Seed to allow for reproducibility.}

\item{register}{Setting for registering the parallel backend with the "foreach" package. To start parallel execution of R code on machine with multiple cores, "NULL" value needs to be assigned to this parameter.}
}
\value{
The function returns an object of "betaclust" class which contains the following values:
\itemize{
\item information_criterion - the information criterion used to select the optimal model.
\item ic_output - this stores the information criterion value calculated for each model.
\item optimal_model - the model selected as optimal.
\item function_call - the parameters passed as arguments to the function betaclust.
\item K - the number of clusters identified using the beta mixture models.
\item C - the number of CpG sites analysed using the beta mixture models.
\item N - the number of patients analysed using the beta mixture models.
\item R - the number of samples analysed using the beta mixture models.
\item optimal_model_results - this contains information from the optimal model. Specifically,
   \itemize{
   \item cluster_size - the total number of CpG sites identified in each of the K clusters.
   \item llk - a vector containing the log-likelihood value at each step of the EM algorithm.
   \item data - this contains the methylation dataset along with the cluster label for each CpG site.
   \item alpha - this contains the first shape parameter for the beta mixture model.
   \item delta - this contains the second shape parameter for the beta mixture model.
   \item tau - the proportion of CpG sites in each cluster.
   \item z - a matrix containing the probability for each CpG site of belonging to each of the \eqn{K} clusters.
   \item uncertainty - the uncertainty of each CpG site's clustering.
   \item thresholds - threshold points calculated for the K.. or the KN. model
   }
}
}
\description{
A family of model based clustering techniques to identify methylation states in beta valued DNA methylation data.
}
\details{
This is a wrapper function which can be used to fit all three models (K.., KN., K.R) together.
The K.. and KN. models are used to analyse a single DNA sample (\eqn{R = 1}) and cluster the \eqn{C} CpG sites into the \eqn{K} clusters which represent the different methylation states in a DNA sample. As each CpG site can belong to either of the \eqn{M=3} methylation states (hypomethylation, hemimethylation and hypermethylation), the default value for \eqn{K=M=3}.
The thresholds between methylation states are objectively inferred from the clustering solution.
The K.R model is used to analyse \eqn{R} independent samples collected from \eqn{N} patients, where each sample contains \eqn{C} CpG sites, and cluster
the dataset into \eqn{K=M^R} clusters to identify the differentially methylated CpG (DMC) sites between the \eqn{R} DNA samples.
}
\examples{
\dontrun{
data(pca.methylation.data)
my.seed = 190
M = 3
patients = 4
samples = 2
data_output = betaclust(pca.methylation.data[,2:9],M,patients,samples,
            model_names = c("K..","KN.","K.R"),model_selection = "BIC",seed = my.seed)
}


}
\references{
{Silva, R., Moran, B., Russell, N.M., Fahey, C., Vlajnic, T., Manecksha, R.P., Finn, S.P., Brennan, D.J., Gallagher, W.M., Perry, A.S.: Evaluating liquid biopsies for methylomic profiling of prostate cancer. Epigenetics 15(6-7), 715-727 (2020). doi:10.1080/15592294.2020.1712876.}

{Majumdar, K., Silva, R., Perry, A.S., Watson, R.W., Murphy, T.B., Gormley, I.C.: betaclust: a family of mixture models for beta valued DNA methylation data.}

{Microsoft, Weston, S. (2022): foreach: Provides Foreach Looping Construct. R package version 1.5.2. https://CRAN.R-project.org/package=foreach.}
}
\seealso{
\code{\link{beta_k}}

\code{\link{beta_kn}}

\code{\link{beta_kr}}

\code{\link{pca.methylation.data}}

\code{\link{plot.betaclust}}

\code{\link{summary.betaclust}}

\code{\link{threshold}}
}
