% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaclust.R
\name{betaclust}
\alias{betaclust}
\title{The betaclust wrapper function}
\usage{
betaclust(
  data,
  K = 3,
  patients,
  samples,
  model_names = "C..",
  model_selection = "BIC",
  seed,
  register = NULL
)
}
\arguments{
\item{data}{Methylation values for \eqn{C} CpG sites from \eqn{R} samples collected from \eqn{N} patients.}

\item{K}{Number of methylation profiles to be identified.}

\item{patients}{Number of patients in the study.}

\item{samples}{Number of samples collected from each patient for study.}

\item{model_names}{Models to run from the set of models, C.., CN. and C.R, default = C.. . See details.}

\item{model_selection}{Information criterion used for model selection. Options are AIC/BIC/ICL/default=BIC.}

\item{seed}{Seed to allow for reproducibility.}

\item{register}{Setting for registering the parallel backend with the "foreach" package. To start parallel execution of R code on machine with multiple cores, "NULL" value needs to be assigned to this parameter.}
}
\value{
The function returns an object of "betaclust" class which contains the following values:
\itemize{
\item Information_criterion - the information criterion used to select the optimal model.
\item ic_output - this stores the information criterion value calculated for each model.
\item optimal_model - the model selected as optimal.
\item function_call - the parameters passed as arguments to the function betaclust.
\item C - the number of CpG sites analysed using the beta mixture models.
\item N - the number of patients analysed using the beta mixture models.
\item R - the numder of samples analysed using the beta mixture models.
\item optimal_model_results - this contains information from the optimal model. Specifically,
   \itemize{
   \item cluster_size - the total number of CpG sites identified in each cluster.
   \item llk - a vector containing the log-likelihood value at each step of the EM algorithm.
   \item data - this contains the methylation dataset along with the cluster label for each CpG site.
   \item alpha - this contains the shape parameter 1 for the beta mixture model.
   \item delta - this contains the shape parameter 2 for the beta mixture model.
   \item tau - the proportion of CpG sites in each cluster.
   \item z - a matrix containing the probability for each CpG site of belonging to each of the \eqn{K} clusters.
   \item uncertainty - the uncertainty of each CpG site's clustering.
   }
}
}
\description{
A family of model based clustering techniques to identify methylation profiles in beta valued DNA methylation data.
}
\details{
This is a wrapper function which can be used to fit all three models (C.., CN., C.R) together.
The C.. and CN. models are used to analyse a single DNA sample (\eqn{R = 1}) and cluster the \eqn{C} CpG sites into the \eqn{K} methylation profiles. As each CpG site can belong to either of the \eqn{M=3} methylation profiles (hypomethylation, hemimethylation and hypermethylation), the default value for \eqn{K=M=3}.
The thresholds between methylation profiles can be objectively identified from the clustering solution.
The C.R model is used to analyse \eqn{R} independent samples collected from \eqn{N} patients, where each sample contains \eqn{C} CpG sites, and cluster
the dataset to identify the differentially methylated CpG sites between the \eqn{R} DNA samples.
}
\examples{
\dontrun{
data(pca.methylation.data)
my.seed=190
K=3
patients=4
samples=2
data_output=betaclust(pca.methylation.data[,2:9],K,patients,samples,
            model_names=c("C..","CN.","C.R"),model_selection="BIC",seed=my.seed)
}


}
\references{
{Silva, R., Moran, B., Russell, N.M., Fahey, C., Vlajnic, T., Manecksha, R.P., Finn, S.P., Brennan, D.J., Gallagher, W.M., Perry, A.S.: Evaluating liquid biopsies for methylomic profiling of prostate cancer. Epigenetics 15(6-7), 715-727 (2020). doi:10.1080/15592294.2020.1712876.}

{Majumdar, K., Silva, R., Perry, A.S., Watson, R.W., Murphy, T.B., Gormley, I.C.: betaclust: a family of mixture models for beta valued DNA methylation data.}
}
\seealso{
\code{\link{beta_c}}

\code{\link{beta_cn}}

\code{\link{beta_cr}}

\code{\link{pca.methylation.data}}

\code{\link{plot.betaclust}}

\code{\link{summary.betaclust}}
}
