% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_aic.R
\name{em_aic}
\alias{em_aic}
\title{AIC to compare the methods}
\usage{
em_aic(llk, C, K, patients = 4, samples = 1, mixm = "C..")
}
\arguments{
\item{llk}{log-likelihood value}

\item{C}{number of CpG sites}

\item{K}{number of clusters}

\item{patients}{number of patients}

\item{samples}{no. of samples}

\item{mixm}{mixture model (method=c("C..","CN.","C.R"))}
}
\value{
aic
}
\description{
AIC to compare the methods
}
