% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_aic.R
\name{em_aic}
\alias{em_aic}
\title{Akaike Information Criterion}
\usage{
em_aic(llk, C, M, N, R, model_name = "K..")
}
\arguments{
\item{llk}{log-likelihood value.}

\item{C}{number of CpG sites.}

\item{M}{number of methylation states identified in a DNA sample.}

\item{N}{number of patients.}

\item{R}{number of DNA samples collected from each patient.}

\item{model_name}{fitted mixture model (model_name = c("K..","KN.","K.R")).}
}
\value{
The AIC value for the selected model.
}
\description{
Compute the AIC value for the optimal model.
}
\details{
Computes the AIC for a specified model given the log-likelihood, the dimension of the data, and the model specification.
}
\seealso{
\code{\link{em_bic}}

\code{\link{em_icl}}
}
