% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_bic.R
\name{em_bic}
\alias{em_bic}
\title{BIC to compare the methods}
\usage{
em_bic(llk, C, K, patients = 4, samples = 1, mixm = "C..")
}
\arguments{
\item{llk}{log-likelihood value}

\item{C}{number of CpG sites}

\item{K}{number of clusters}

\item{patients}{number of patients}

\item{samples}{no. of samples}

\item{mixm}{mixture model (method=c("C..","CN.","C.R"))}
}
\value{
bic
}
\description{
BIC to compare the methods
}
