% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_bic.R
\name{em_bic}
\alias{em_bic}
\title{Bayesian Information Criterion}
\usage{
em_bic(llk, C, K, patients = 4, samples = 1, model_name = "C..")
}
\arguments{
\item{llk}{log-likelihood value.}

\item{C}{number of CpG sites.}

\item{K}{number of methylation profiles identified.}

\item{patients}{number of patients.}

\item{samples}{number of DNA samples collected from each patient.}

\item{model_name}{fitted mixture model (method=c("C..","CN.","C.R")).}
}
\value{
The BIC value for the selected model.
}
\description{
Compute the BIC value for the optimal model.
}
\details{
Computes the BIC for a specified model given the log-likelihood, the dimension of the data, and the model specification.
}
\seealso{
\code{\link{em_aic}}

\code{\link{em_icl}}
}
