% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_icl.R
\name{em_icl}
\alias{em_icl}
\title{ICL to compare the methods}
\usage{
em_icl(llk, C, K, patients = 4, samples = 1, model_names = "C..", z)
}
\arguments{
\item{llk}{log-likelihood value}

\item{C}{number of CpG sites}

\item{K}{number of clusters}

\item{patients}{number of patients}

\item{samples}{no. of samples}

\item{model_names}{mixture model (method=c("C..","CN.","C.R"))}

\item{z}{z matrix for each output}
}
\value{
icl
}
\description{
ICL to compare the methods
}
