% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_icl.R
\name{em_icl}
\alias{em_icl}
\title{Integrated Complete-data Likelihood (ICL) Criterion}
\usage{
em_icl(llk, C, K, patients = 4, samples = 1, model_names = "C..", z)
}
\arguments{
\item{llk}{log-likelihood value}

\item{C}{number of CpG sites}

\item{K}{number of clusters}

\item{patients}{number of patients}

\item{samples}{no. of samples}

\item{model_names}{mixture model (method=c("C..","CN.","C.R"))}

\item{z}{z matrix for each output}
}
\value{
The ICL value for the selected model.
}
\description{
The ICL value used to select the optimal model.
}
\details{
Computes the ICL for the beta mixture models given the loglikelihood,
the dimension of the data, and the mixture model names. This criterion penalises the BIC by including the entropy term favouring the
well separated clusters.
}
\seealso{
\code{\link{em_aic}}

\code{\link{em_bic}}
}
