% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_icl.R
\name{em_icl}
\alias{em_icl}
\title{Integrated Complete-data Likelihood (ICL) Criterion}
\usage{
em_icl(llk, C, K, patients = 4, samples = 1, model_name = "C..", z)
}
\arguments{
\item{llk}{log-likelihood value.}

\item{C}{number of CpG sites.}

\item{K}{number of methylation profiles to be identified.}

\item{patients}{number of patients in the study.}

\item{samples}{number of samples collected from each patient for study.}

\item{model_name}{mixture model (method=c("C..","CN.","C.R")).}

\item{z}{z matrix used for computing the complete-data log-likelihood function.}
}
\value{
The ICL value for the selected model.
}
\description{
Compute the ICL value for the optimal model.
}
\details{
Computes the ICL for a specified model given the log-likelihood, the dimension of the data, and the model specification. This criterion penalises the BIC by including the entropy term favouring the
well separated clusters.
}
\seealso{
\code{\link{em_aic}}

\code{\link{em_bic}}
}
