% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.betaclust.R
\name{plot.betaclust}
\alias{plot.betaclust}
\title{Plots for visualizing the betaclust class object}
\usage{
\method{plot}{betaclust}(object, what = "density", plot_type = "ggplot", scale_param = "free_y")
}
\arguments{
\item{object}{betaclust object}

\item{what}{The different plots that can be obtained from the object (default="density") (what=c("density","uncertainty","InformationCriterion"))}

\item{plot_type}{The plot type to be displayed (default="ggplot")(plot_type="ggplot" or"plotly")}

\item{scale_param}{The axis that needs to be fixed or not for facet plot (scales=c("free_y","free_x","free"), default is "free_y")}
}
\description{
This function helps visualise the clustering solution by plotting the density estimates, the uncertainty and the information criterion.
}
\details{
The density estimates of the clustering solution of the optimal model can be
plotted by passing the parameter what="density" in the function. Apart from static plots interactive plots can also be plotted using the
parameter plot_type = "plotly". The uncertainty in the clustering soluting can be plotted using what="uncertainty".
The information criterion values for all models can be plotted using what="InformationCriterion" for selecting the optimal model.
}
\seealso{
\code{\link{betaclust}}
}
