% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.betaclust.R
\name{plot.betaclust}
\alias{plot.betaclust}
\title{Plots for betaclust object}
\usage{
\method{plot}{betaclust}(object, what = "density", plot_type = "ggplot")
}
\arguments{
\item{object}{betaclust object}

\item{what}{The different plots that can be obtained from the object (default="density") (what=c("density","uncertainty","InformationCriterion"))}

\item{plot_type}{The plot type to be displayed (default="ggplot")(plot_type="ggplot" or"plotly")}
}
\description{
Plots for betaclust object
}
