% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.betaclust.R
\name{plot.betaclust}
\alias{plot.betaclust}
\title{Plots for visualizing the betaclust class object}
\usage{
\method{plot}{betaclust}(
  object,
  what = "fitted density",
  plot_type = "ggplot",
  sample_name = NULL,
  title = NULL,
  patient_number = 1,
  threshold = FALSE,
  scale_param = "free_y"
)
}
\arguments{
\item{object}{A betaclust object.}

\item{what}{The different plots that can be obtained are either "fitted density","kernel density","uncertainty" or "information criterion". (default = "fitted density").}

\item{plot_type}{The plot type to be displayed are either "ggplot" or "plotly". (default = "ggplot").}

\item{sample_name}{The names of DNA samples in the dataset analysed by the K.R model. If no value is passed then default values of sample names, e.g. Sample 1, Sample 2, etc are used as legend text (default = NULL).}

\item{title}{The title that the user wants to display. If no title is to be displayed the default is "NULL".}

\item{patient_number}{The column number representing the patient in the patient-wise ordered dataset selected for visualizing the clustering solution of the K.. or KN. model (default = 1).}

\item{threshold}{The "TRUE" option displays the threshold points in the graph for the K.. and the KN. model (default = "FALSE").}

\item{scale_param}{The axis that needs to be fixed for density estimates plot for visualizing the K.R clustering solution are either "free_y","free_x" or "free" (default = "free_y").}
}
\description{
Visualise a \code{\link[betaclust:betaclust]{betaclust}} clustering solution by plotting the fitted and kernel density estimates, the uncertainty and the information criterion.
}
\details{
The fitted density estimates can be visualized under the optimal clustering solution by specifying what = "fitted density" and kernel density estimates under the optimal clustering solution by specifying what = "kernel density". The threshold inferred can be visualized by specifying threshold = TRUE.
 The KN. model calculates different pairs of threshold points for each patient as the shape parameters are allowed to vary for each patient. So the patient for whom the threshold needs to be displayed can be specified by inputting the column number representing the patient in the patient-wise ordered dataset in the parameter patient_number.
 Interactive plots can also be produced using plot_type = "plotly". The uncertainty in the clustering solution can be plotted using what = "uncertainty".
 The information criterion values for all fitted models can be plotted using what = "information criterion".
}
\examples{
\dontrun{
M = 3
N = 4
R = 2
data_output = betaclust(pca.methylation.data[,2:9],M,N,R,
            model_names = c("K..","KN.","K.R"),model_selection = "BIC",seed = my.seed)
plot(data_output,what = "fitted density",plot_type = "ggplot")}
}
\seealso{
\code{\link{betaclust}}
}
