% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.betaclust.R
\name{print.summary.betaclust}
\alias{print.summary.betaclust}
\title{ICL to compare the methods}
\usage{
\method{print}{summary.betaclust}(x, digits = getOption("digits"), ...)
}
\arguments{
\item{x}{summary of betaclust object}
}
\description{
ICL to compare the methods
}
