% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.betaclust.R
\name{summary.betaclust}
\alias{summary.betaclust}
\title{Calculating summary of the betaclust object}
\usage{
\method{summary}{betaclust}(object)
}
\arguments{
\item{x}{betaclust object}
}
\value{
obj
}
\description{
Calculating summary of the betaclust object
}
