% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.betaclust.R
\name{summary.betaclust}
\alias{summary.betaclust}
\title{ICL to compare the methods}
\usage{
\method{summary}{betaclust}(object)
}
\arguments{
\item{x}{betaclust object}
}
\value{
obj
}
\description{
ICL to compare the methods
}
