% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.betaclust.R
\name{summary.betaclust}
\alias{summary.betaclust}
\title{Summarizing the beta mixture model fits}
\usage{
\method{summary}{betaclust}(object)
}
\arguments{
\item{object}{A betaclust object.}
}
\value{
An object of class "summary.betaclust which contains the following list of values:
\itemize{
\item C - the number of CpG sites analysed using the beta mixture models.
\item N - the number of patients analysed using the beta mixture models.
\item R - the numder of samples analysed using the beta mixture models.
\item K - the number of methylation profiles identified.
\item modelName - the optimal model selected.
\item loglik - the log-likelihood value for the selected optimal model.
\item information_criterion - the information criterion used to select the optimal model.
\item ic_output - this stores the information criterion value calculated for each model.
\item classification - the total number of CpG sites identified in each cluster.
\item prop_data - the proportion of CpG sites identified in each cluster}
}
\description{
Summary method for a "betaclust" object containing the results under the optimal model selected.
}
\examples{
\dontrun{
data_output=betaclust(pca.methylation.data[,2:9],K,patients,samples,
            model_names=c("C..","CN.","C.R"),model_selection="BIC",seed=my.seed)
summary(data_output)}
}
\seealso{
\code{\link{betaclust}}
}
